Correct Sass test failure
Tests began to fail as we use dart sass for True Sass tests now in
origami-build-tools v10. We assumed updating to use dart sass for
True tests was not breaking as we already test components compile
with dart sass but we didn't think about the small changes in
output that would impact test assertions.